requirements.txt
streamlit==1.30.0
langchain
python-dotenv
langchain-google-genai
youtube-transcript-api
sentence-transformers
faiss-cpu
faiss-gpu